+++ src/main/java/duke/Duke.java
@@ -1,6 +1,11 @@
-import java.time.format.DateTimeParseException;
-import java.util.ArrayList;
+package duke;
+
 import java.util.Scanner;
+
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.util.Parser;
+import duke.util.Ui;
 
 /**
  * Represents a chatbot
