+++ src/main/java/duke/storage/Storage.java
@@ -1,5 +1,13 @@
+package duke.storage;
+
 import java.util.ArrayList;
 import java.util.List;
+
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.Todo;
+import duke.task.Deadline;
+
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
